+++ lib/l10n/app_localizations_fr.dart
@@ -685,6 +685,10 @@ class AppLocalizationsFr extends AppLocalizations {
       'iOS limitation: Empty folders cannot be selected. Choose a folder with at least one file.';
 
   @override
+  String get setupIcloudNotSupported =>
+      'iCloud Drive is not supported. Please use the app Documents folder.';
+
+  @override
   String get setupDownloadInFlac => 'Download Spotify tracks in FLAC';
 
   @override